Add support for meta data
We add support for adding meta data to events by providing a store where
you can check which event you have received and based on that provide
additional information that you don't want to store with your event.

Metadata handling is optional, and as such this does not break backwards
compatibility.